Allow `__init__` kwargs to be assigned as object attributes in the design (#95)
* refactor: Test case refactoring

Refactored builder unit tests to make the tests clearer and easier to define. This should make it easier to have tests for different versions of Nautobot

* fix: Fixed error message construction

When constructing an error string there was an issue where another exception could render the string unreadable. This fix wraps the errant code with try/except to continue producing a usable error string.

* feat: Added the ability to mark YAML tests to be skipped

* test: Initial tests for Nautobot 2.0

* feat: Allow kwargs as design object attributes.