fix: Fixes code injection vulnerability
This change provides a way to prevent strings being evaluated as jinja2 templates. The change also sets the `Context` to not evaluate strings as templates. Since user input (script vars) are passed into the Context this will mean that they are not evaluated, but other strings provided in context files and context classes will still be evaluated is usual.

Fixes #258